Create TIC TAC TOE GAME
Tic-tac-toe (American English), noughts and crosses (Commonwealth English), or Xs and Os (Canadian or Irish English) is a paper-and-pencil game for two players who take turns marking the spaces in a three-by-three grid, one with Xs and the other with Os. A player wins when they mark all three spaces of a row, column, or diagonal of the grid, where upon they traditionally draw a line through those three marks to indicate the win. It is a solved game, with a forced draw assuming best play from both players.